requirements.txt
absl-py==2.1.0
aiohttp
aiobotocore
aioitertools
alabaster
altair
appdirs==1.4.4
appnope
argon2-cffi
argon2-cffi-bindings
astroid
astropy
astunparse==1.6.3
async-lru
atomicwrites==1.4.0
attrs
Automat
autopep8
Babel
bcrypt
beautifulsoup4
black
bleach
bokeh
boto3
botocore
Bottleneck
Brotli
cachetools
certifi
cffi
chardet
click
cloudpickle
colorama
colorcet
cryptography
cssselect
cycler
Cython
dask
datashader
decorator
defusedxml
diff-match-patch
dill
distributed
distro
docstring-to-markdown
docutils
entrypoints
et-xmlfile
executing
fastjsonschema
filelock
flake8
Flask
flatbuffers==24.12.23
fonttools
frozendict
frozenlist
fsspec
gast==0.6.0
gensim
grpcio==1.68.1
h11
h5py
HeapDict
holoviews
httpcore
httpx
hvplot
hyperlink
idna
imagecodecs
imageio
imagesize
imbalanced-learn
importlib-metadata
iniconfig
intake
intervaltree
ipykernel
ipython
ipython-genutils
ipywidgets
isort
itemadapter
itemloaders
itsdangerous
jedi
jellyfish
Jinja2
jmespath
joblib
json5
jsonpatch
jsonpointer==2.1
jsonschema
jupyter
jupyter-console
jupyter-client
jupyter-core
jupyter-server
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab-server
keras==3.7.0
kiwisolver
libclang==18.1.1
llvmlite
lmdb
lz4
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdit-py-plugins
mistune
ml-dtypes==0.4.1
more-itertools
mpmath
msgpack
multidict
multipledispatch
mypy
mypy-extensions
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
numba
numexpr
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
overrides
packaging
pandas==2.2.2
panel
param
parso
partd
pathspec
patsy
pexpect
pickleshare
Pillow
Pint
platformdirs
plotly
pluggy
ply
prometheus-client
prompt-toolkit
protobuf==3.20.3
psutil
ptyprocess
pure-eval
pyarrow
pyasn1
pyasn1-modules==0.2.8
pycodestyle
pycosat
pycparser
pyct
pycurl
pydantic
pydantic-core
pydeck
PyDispatcher
pydocstyle
pyerfa
pyflakes
Pygments
PyJWT
pylint
pyls-spyder==0.4.0
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-CoreServices
pyobjc-framework-FSEvents
pyodbc
pyOpenSSL
pyparsing
PyQt5==5.15.10
PyQt5-sip
PyQtWebEngine==5.15.6
PySocks
pytest
python-dateutil
python-dotenv
python-json-logger
python-lsp-black
python-lsp-jsonrpc
python-lsp-server
python-slugify
python-snappy
pytoolconfig
pytz
pywavelets
PyYAML
pyzmq
QDarkStyle
qstylizer
QtAwesome
qtconsole
QtPy
queuelib
regex
requests
requests-file
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rope
ruamel.yaml
s3fs
scikit-image
scikit-learn
scipy==1.13.1
Scrapy
seaborn
semver
Send2Trash
setuptools==69.5.1
sip
six
smart-open
smmap
sniffio
snowballstemmer
sortedcontainers
soupsieve
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
SQLAlchemy
stack-data
statsmodels
streamlit
sympy
tables
tabulate
tblib
tenacity
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
terminado
text-unidecode
textdistance
threadpoolctl
three-merge
tifffile
tinycss2
tldextract
toml
tomli
tomlkit
toolz
tornado
tqdm
traitlets
truststore
ttkbootstrap==1.10.1
Twisted
typing_extensions
tzdata
uc-micro-py
ujson
unicodedata2
Unidecode
urllib3
w3lib
watchdog
wcwidth
webencodings
websocket-client
Werkzeug
whatthepatch
wheel==0.43.0
widgetsnbextension
wrapt
wurlitzer
xarray
xlwings
xyzservices
yapf
yarl
zict
zipp
zope.interface
zstandard
